requirements.txt
cassandra-driver==3.24.0

